client3_1.py
We were able to create a prices dictionary to store the stock prices. Think of a dictionary
as a key-value store wherein you can specify a key and be able to retrieve a
value. In our case, they key was the stock name and the value was the price.
We then used this prices dictionary at the end to pass in the right values in
the getRatio function.