client_test_1.py
In the case of this task, the methods we want to test are getDataPoint and
getRatio. We want to cover these with tests so that we ensure these
methods are returning the correct output as individual methods thereby
ensuring us that when we use them together in the main client application,
things will work fine.
In unit testing, we always follow the pattern, Build-Act-Assert. Assertions should mostly be used once and be the strongest.